DHFPROD-5943: Renaming the connector to Spark Connector
