--- marklogic-data-hub-glue-connector/src/main/java/com/marklogic/hub/cloud/aws/glue/Writer/MarkLogicWriteDataSource.java
@@ -1,34 +1,0 @@
-/*
- * Copyright 2020 MarkLogic Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.marklogic.hub.cloud.aws.glue.Writer;
-
-import com.marklogic.client.ext.helper.LoggingObject;
-import org.apache.spark.sql.SaveMode;
-import org.apache.spark.sql.sources.v2.DataSourceOptions;
-import org.apache.spark.sql.sources.v2.WriteSupport;
-import org.apache.spark.sql.sources.v2.writer.DataSourceWriter;
-import org.apache.spark.sql.types.StructType;
-
-import java.util.Optional;
-
-public class MarkLogicWriteDataSource extends LoggingObject implements WriteSupport {
-
-    @Override
-    public Optional<DataSourceWriter> createWriter(String writeUUID, StructType schema, SaveMode mode, DataSourceOptions options) {
-        logger.info("Creating MarkLogicWriter");
-        return Optional.of(new MarkLogicWriter(options.asMap(), schema));
-    }
-}
