+++ marklogic-data-hub-spark-connector/gradle.properties
