+++ marklogic-data-hub-spark-connector/src/main/java/com/marklogic/hub/spark/sql/sources/v2/writer/HubDataWriterFactory.java
@@ -13,11 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.marklogic.hub.cloud.aws.glue.Writer;
+package com.marklogic.hub.spark.sql.sources.v2.writer;
 
 import com.marklogic.client.ext.helper.LoggingObject;
 import com.marklogic.hub.HubClient;
 import com.marklogic.hub.impl.HubConfigImpl;
+import com.marklogic.hub.spark.sql.sources.v2.writer.HubDataWriter;
 import com.marklogic.mgmt.util.SimplePropertySource;
 import org.apache.spark.sql.catalyst.InternalRow;
 import org.apache.spark.sql.sources.v2.writer.DataWriter;
@@ -27,7 +28,7 @@ import org.apache.spark.sql.types.StructType;
 import java.util.Map;
 import java.util.Properties;
 
-public class MarkLogicDataWriterFactory extends LoggingObject implements DataWriterFactory<InternalRow> {
+public class HubDataWriterFactory extends LoggingObject implements DataWriterFactory<InternalRow> {
 
     private StructType schema;
     private HubClient hubClient;
@@ -39,7 +40,7 @@ public class MarkLogicDataWriterFactory extends LoggingObject implements DataWri
      *               HubClient for communicating with MarkLogic.
      * @param schema
      */
-    public MarkLogicDataWriterFactory(Map<String, String> params, StructType schema) {
+    public HubDataWriterFactory(Map<String, String> params, StructType schema) {
         this.params = params;
         this.schema = schema;
 
@@ -57,6 +58,6 @@ public class MarkLogicDataWriterFactory extends LoggingObject implements DataWri
         if (logger.isDebugEnabled()) {
             logger.debug("Creating DataWriter with taskId: " + taskId);
         }
-        return new MarkLogicDataWriter(hubClient, taskId, schema, params);
+        return new HubDataWriter(hubClient, taskId, schema, params);
     }
 }
