+++ marklogic-data-hub-spark-connector/src/test/resources/log4j.properties
