+++ settings.gradle
@@ -3,6 +3,6 @@ include ':marklogic-data-hub'
 include ':ml-data-hub-plugin'
 include ':web'
 include ':marklogic-data-hub-central'
-include ':marklogic-data-hub-glue-connector'
+include ':marklogic-data-hub-spark-connector'
 
 project(':ml-data-hub-plugin').name = 'ml-data-hub'
